requirements.txt
fastapi
uvicorn
jinja2
pdfplumber
nltk
python-multipart
python-dotenv
requests
